Add scan engine.
Why: This commit adds the ScannerModule, the CoreScanner, and the
ScanEngineModule and ScanEngineController. It also adds tests for all of
the components. This structure allows us to add new scans in the
ScannersModule without affecting the ScanEngineModules controller.

How: The directories to explore here include `apps/scanner/src/scan-engine`
and `apps/scanner/src/scanner`. The `core` scanner is in the latter
directory and is the base scanner for the project. Additionally, I am
using Nest.js for it's IoC container, and Jest for testing.

Tags: scan-engine, scanners, core scanner, nest, jest, Nest.js, testing