Add local development setup.
Why: In order to test out the message queue architecture, add a
Docker-based local development setup. This splits the application into
the API and the Scanner. The app is now ready to add the message
queueing system to the scanner application.

How: Add Dockerfiles, docker-compose.yml and documentation. The scanner
Dockerfile is worth taking a look at as it required some interesting
things that I hadn't played around with before. Additionally, I am
adding an .nvmrc so that we can have a pinned version of node.

Tags: Docker, docker-compose, redis, local development, documentation