Add Message Queue, Producer, and Consumers
Why: This adds a cron-triggered message-queuing system to the app.
This system is used to add URLs that need to be scanned to a queue, along with
some metadata. Then the scanner system pulls the work off and processes it.

Additionally, this adds the queue and the services to Docker for local development.

Also, it adds the default Nest.js ConfigModule which leverages dotenv.

How: The queue is a Nest.js wrapper around Bull. Bull is a
high-performance Redis-backed queue manager in Node.js. This system also
adds a Cron task (that currently runs once a minute).

Tags: Docker, queues, bull, redis