Add infrastructure (#1)
* Add local development setup.

Why: In order to test out the message queue architecture, add a
Docker-based local development setup. This splits the application into
the API and the Scanner. The app is now ready to add the message
queueing system to the scanner application.

How: Add Dockerfiles, docker-compose.yml and documentation. The scanner
Dockerfile is worth taking a look at as it required some interesting
things that I hadn't played around with before. Additionally, I am
adding an .nvmrc so that we can have a pinned version of node.

Tags: Docker, docker-compose, redis, local development, documentation

* Add Message Queue, Producer, and Consumers

Why: This adds a cron-triggered message-queuing system to the app.
This system is used to add URLs that need to be scanned to a queue, along with
some metadata. Then the scanner system pulls the work off and processes it.

Additionally, this adds the queue and the services to Docker for local development.

Also, it adds the default Nest.js ConfigModule which leverages dotenv.

How: The queue is a Nest.js wrapper around Bull. Bull is a
high-performance Redis-backed queue manager in Node.js. This system also
adds a Cron task (that currently runs once a minute).

Tags: Docker, queues, bull, redis

* Add database and add libs directory for db and message queue.

Why: Using postgres for the datastore. I'm using TypeOrm because it
intergrates nicely with Nest.js. Moved the message queue into it's own
module within the libs directory.

Tags: database, postgres, typeorm, message queue

* Add CoreResult.

Add the the CoreResult concept to the system. This gives us the ability
to persist results from the CoreScanner.