Add API to cloud.gov
Why: This adds the site-scanning-api to Cloud.gov.

How: Add a deploy-cloudgov.sh script which (currently) only creates the
services necessary. Note that there is some logic in the script that
checks if Postgres is created, but I just read that this logic is
actually incorrect! I'll update in the next PR.

Add a manifest.yml to specify the api name and how
to run it -- this also binds the postgres RDS instance to the
application.

Lastly, update the database configuration so that we pull in the
credentials using the ConfigService.

tags: postgres, cloud.gov, config, api