Add message queue to cloud.gov (#3)
Why: This adds the message queue implementation into Cloud.gov infrastructure. There's a few things going on here:

I refactored the configuration for the message queue to get the credentials from VCAP_SERVICES or locally if VCAP_SERVICES does not exist.
I added to the manifest.yml the site-scanning-producer and the site-scanning-consumer applications. The consumer deserves a bit of explanation: Because the consumer needs to run puppeteer which uses Chromium we have to use two buildpacks the apt and the nodejs buildpack. apt.yml contains all of the apt packages that we need. For now, we are using the --no-sandbox argument, but I will be removing this shortly.