Add ingest service and CLI.
Why: The IngestService can be used for any kind of data ingest into the
system, but right now it used to add target urls for the scanner.

The CLI is also general purpose, but right now it is being used to call
the IngestService and add target urls to the scanner.

How: Add an `ingest` library and a `cli` app. The app uses the library.