Add initial Target URL generation capability (#17)
Why: We need a way to repeatably add target urls for the scanner to scan. This commit adds a few things in service of that goal:

An IngestService that can be used for any kind of data ingest into the
system. In this commit it is used to add target urls to the Website table which the scanner consumes.

A general-purpose CLI for the site-scanner. In this commit it is being used to call
the IngestService and add target urls to the scanner.

Updates to the Website model to include new fields from the GSA dataset.

A fix to the CoreScanner to prepend "https" to the target url.

Tags: target urls, core scanner, ingest, cli, docs