Add fields to CoreScanner (#19)
Why: This adds several of the Scanner V2 Fields to the CoreScanner. Additionally, it makes changes to the database schema and to the API to accommodate the new fields. Of note in the database is a switch to a one-to-one mapping between CoreResults and Websites. Since we are treating this more or less like a daily cache, this is a useful relation.

The API introduces a mapper between the CoreResult repository object and the JSON schema that that we want to display. I'm not 100% sure I like this design just yet, but it's working for now.

Lastly, I moved the Core Input/Output DTOs into the core scanner itself.

Tags: API, core scanner, database, schema change, one to one, mapper