Add Scan Status and Scan Date. Create CoreResult and Website Entities. (#21)
Why: This adds the `scan_status` and `scan_date` fields to the returned fields. Additionally, it does a fairly large refactor of the `CoreResult` model. This was a good time for the refactor, because I had to make fields nullable in `CoreResult` in order to support the `scan_status` work.

The refactor removes the output DTOs and centralizes around the `CoreResult` and `Website` entities. This was necessary because there was way too much field mapping going on between the DTOs and the entity. The tests and serializers are much cleaner now.

How: Moving the `Website` and `CoreResult` into the new `entities` directory. Adding serializer decorators to the fields so that we can call them with the `class-transformer` library to get nicely serialized versions of them. 

Tags: refactor, core result, fields.