Add redirect test. (#22)
Why: This tests that a target_url returns a 404 when given a page that doesn't exist. The current methodology is just to add a path not_found_test_{uuid} with a random uuid generated.

How: Add an axios get call to the core scanner. Add tests.

Tags: core scanner, not found test, uuid, 404