Add USWDS Scan (#28)
Why: This adds a scanner and entity specifically for USWDS scans. This adapts Domain Scan's USWDS2 scanner. The basic idea behind this scanner is to heuristically determine the presence of USWDS on given website. It checks for the various characteristics that would likely be present on a site using USWDS and weights them accordingly.

How: Add a new library uswds-scanner which implements the scanner interface. Add a new entity uswds-result.entity.ts which captures all of the fields. and various serialization methods.

Tags: uswds, scanner, entity, heuristic