Add uswds version fields. (#33)
Why: This adds the USWDS version fields to the Solutions scanner. This looks in CSS for the first USWDS version it can find (using RegEx) and returns the version. It also adds 20 points to the USWDS Detection heuristic.

How: Add fields to the scanner and the data model. Add tests.

Tags: uswds, semantic version, solutions scanner