Bug and Memory Leak Fixes (#34)
Why: I detected a memory leak in Cloud.gov because I wasn't properly closing the puppeteer pages in the scanners. This fixes that and reorganizes the code to be much more readable. Additionally, the named jobs in the producer/consumer do not appear to working as expected, so I am just running one job (with double the concurrency). This is working fine so far.

How: I added a finally to the try...catch which closes the puppeteer pages. I moved the actual creation of the result objects to a different method and I'm keeping all of the network logic in the scan methods. Removed the solutions specific producer/consumer logic.

Tags: memory leak, bug fixes, concurrency, scanners, producer