Ignore incomplete chain errors from axios. (#37)
Why: There is inconsistency in the data, because we are using Axios for the 404 test and Axios throws an SSL error when it can't verify a certificate chain. For now, disable checking that. Add a test for a website that has an incomplete chain.

Tags: axios, 404 test, testing, SSL, certificate chain