Add DAP Scan (#39)
Why: This adds a scan that looks for the presence of DAP on a webpage. It looks for a few things: 1) The presence of an outbound request url or post body, that has DAPs unique Google Analytics UA Id 2) It looks for the parameters on an outbound request to DAP and includes them as an object. Closes 18F/site-scanning#797

How: Add methods and fields to the SolutionsScanner. Add e2e test. Note the (perhaps kludgy) @Transform decorator added to the SolutionsResult entity. This transforms the dap parameters field to an object on serialization to a POJO.

Tags: DAP, digital analytics program, solutions scanner, testing