Fix Date Issue: Pt. 2 (#62)
Why: There was a bug in the way I was saving both CoreResults and SolutionsResults. This basically meant that the original updated date would always be persisted on updates to the database. This change uses the insert and update methods of the Typeorm repository so that the existing object is updated (or the new object is created).

How: Change the core-results-service and solutions-results-service. Update the tests.

Tags: bugfix, typeorm, updated, core-result, solutions-result