Add Bulk Download Feature (#64)
Why: This feature adds a JSON snapshot of the database to S3. This allows for bulk downloading from S3 (as opposed to through the API). It also archives any existing files and copies them to an /archive directory in S3. There is also a local version set up, that uses minio in place of S3.

How: This feature adds a few things to the codebase. 1) The StorageService which interfaces with S3, 2) The SnapshotService which is responsible for the logic of querying the database, and archival and storage into S3, 3) and the DatetimeService which mostly exists for making testing easier.

The main public API to be concerned with is probably the SnapshotService.weeklySnapshot(). This is used by the TaskService which manages the cron logic. This is currently being called daily (while testing the feature) but has an independent schedule set by the SNAPSHOT_SCHEDULE environment variable.

Closes 18F/site-scanning#824

Tags: s3, archival, storage, snapshot