Add new target url list. (#66)
Why: This adds the 25k target url list to the system -- this necessitated some changes to the IngestService. I also took the opportunity to clean up some issues with how the IngestService dealt with promises and with how the ScanEngineConsumer dealt with the message queue. As this increases the scale of the service by about 25x, I added a lot more logging to various services and event monitoring to the message queue consumer.

A few other things to note: The 25k list has a slightly different schema, so I changed the Website entity schema to match. I added a convenience script for running the ingest task in Cloud.gov.

Tags: ingest, target-url, logging, scan engine