Use to lower instead of toUpper. Fix order of interceptors. (#70)
Why: This is related to the ingest of the 25k urls in which the schema changed slightly from before. The issue is that the previous dataset had all urls in upper case. This change ensures that they're ingested as lower case and queried as lower case. The more sustainable long-term way to do this is to likely use a createQueryBuilder and use something like LOWER(url) = LOWER(:url) in the Typeorm query. But, the tradeoff is losing the expressivity of the repository pattern. Additionally, this change adjusts the order of the Interceptors so that we get a 404 when the target URL doesn't exist.

Tags: bugfix, ingest, api