Add Open API Specification and Cleanup API Validation and Schema (#72)
Why: This adds the OpenAPI/Swagger specification to the API at {api_url}/api or for JSON {api_url}/api-json. It also cleans up the API schema by adding types and takes advantage of some of the features of TypeScript to create some reusable abstractions. A few other small changes, like field name fixes and removing the common directory are also rolled in.

How: This is accomplished by adding the nestjs/swagger library to the compiler options, which allows us to (mostly) autogenerate the Swagger docs without adding a lot of boilerplate to the code. The exception is in the controllers where the return types have to be explicitly decorated onto the functions.

Tags: Swagger, Open API, API, validation

Closes 18F/site-scanning#847