Add API key auth to /websites endpoint. (#126)
What: This adds API key auth to the /websites endpoint. This is meant to be used with api.data.gov which sends an API Key in the X-Secret-Api-Access-Token header, which is then checked on the application side of the API.

How: This pull request creates an ApiKeyGuard which implements the necessary logic. In NestJS, Guards are checked in the request lifecycle, before access to a route is allowed which makes them a natural fit for auth.

This work also adds the ConfigService to the api application which is used to find the API credential that is in turn provided using a Cloud Foundry User Provided Service (aka cf cups). You'll see the updates for the cups in the Cloud Foundry manifest.ymls and in the infrastructure script.

Other changes:
I updated the dependency vulnerabilities from npm audit. This caused a bit of extra work that is really outside of the scope of the feature work. A few things to be aware of:

I had to lose some type information in the unit tests because Typescript could not calculate the Union type of the Repository mocks.

I believe that NestJS deprecated the Logger and replaced it with the ConsoleLogger so the project's LoggerService now extends the ConsoleLogger which has a slightly different API.