Add default target to cloudgov-deploy.sh
Originally, the script would, if passed nothing else, assume the operator
wanted to use manifest.yml regardless of the environment (target); this
looks to see what target is setup for the current session and default to
using that manifest-(target).yml filename.  The default, if nothing is
provided, is manifest-prod.yml which is a symlink to manifest.yml so
there ought to be no changes to existing workflows (i.e., running the
script without an argument still uses the manifest.yml configuration).

Also, the default takes into account the project root so that if the
script is run from a subdirectory, it'll still look for (and, hopefully
find) the correct manifest from the project root.

[Ticket: 136]