Suppress CodeQL warning
When CodeQL first ran, it detected core-scanner.service.ts:184 as being
problematic because it disabled SSL certificate validation.  That's a
good find from a good scan.

However, the context of the containing function is to determine if a
request for a given resource returns an HTTP not-found error (e.g., 404)
when a randomy generated path to a given hostname is there or not.  The
content of that random URL isn't used.  The only thing this function does
is to return whether or not a given `GET` to a random path at the
requested hostname is there or not.  (tangentially, could we replace
`.get()` with `.head()` ?